feat: Update to eslint-plugin-jest 24
BREAKING CHANGE: We have enabled the style set of rules for jest. Most of the rules were enabled manually via overrides, so we just extend the style set now instead. This has turned on two rules: `no-alias-methods` and `prefer-to-contain` both of which are fixable. We also updated to the new set of recommended rules with v24. `no-done-callback` has been set to off as we use done in many places. We also manually enabled `no-large-snapshots`.