feat: Expand list of typescript rules
BREAKING CHANGE: We are expanding your list of typescript rules which requires connecting your eslint config to your tsconfig. See the typescript-eslint documentation https://typescript-eslint.io/linting/typed-linting for how to do this. If your project does not use typescript and you do not want to add a tsconfig file needlessly, disable the following rules in your config: @typescript-eslint/no-confusing-void-expression, @typescript-eslint/no-for-in-array, @typescript-eslint/prefer-includes, @typescript-eslint/prefer-reduce-type-parameter, and @typescript-eslint/prefer-string-starts-ends-with.